+++ ChangeLog.md
@@ -7,6 +7,7 @@ TODO : pour les options commencer à voir si possible de crééer des tab par el
 
 # UNRELEASED
 
+- FIX : Security issue with ticket *1.35.1* - **22/02/2022**
 - NEW : jSignature lib included in externalaccess *1.35.0* - **17/11/2022**
 
 # RELEASE 1.34